Move Activity Attachment Extensions handling into ECMS
