--- webapp/portlet/src/main/webapp/vue-apps/activity-composer-app/components/initComponents.js
@@ -1,3 +1,0 @@
-import ExoActivityComposer  from './ExoActivityComposer.vue';
-
-Vue.component('exo-activity-composer', ExoActivityComposer);
