+++ webapp/portlet/src/main/webapp/vue-apps/activity-composer-app/initComponents.js
@@ -1,0 +1,11 @@
+import ActivityComposer  from './components/ActivityComposer.vue';
+import ActivityComposerDrawer  from './components/ActivityComposerDrawer.vue';
+
+const components = {
+  'activity-composer': ActivityComposer,
+  'activity-composer-drawer': ActivityComposerDrawer,
+};
+
+for (const key in components) {
+  Vue.component(key, components[key]);
+}
